What is PancakeSwap
What is PancakeSwap
PancakeSwap is a decentralized exchange (DEX) on the BSC network that employs an automated market maker mechanism to facilitate token swaps. It was launched in September 2020 by a bunch of anonymous developers. It is also a permissionless DEX which allows anyone to list their tokens on the exchange as long as they create a liquidity pool for the token.

PancakeSwap offers a variety of features for its users such as token swaps, liquidity provision and farming, staking, perpetual trading, lottery, NFT marketplace, launchpad etc. It is one the few DEXs that offers users a more well-rounded Defi experience with access to a variety of financial products in one interface. This sort of user experience is usually native to centralized exchanges that work with huge operational teams and funding. DEXs tend to focus on one or two financial products and stick to them